move job scripts out of resources.
